+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to
 
 - 🐛(docs) run migration 0027 without superuser role
 - 🐛(backend) prevent admins/owners from overwriting other users comments
+- 🐛(y-provider) return empty output when converting empty Yjs document
 
 
 ## [v5.1.0] - 2026-05-11